feat: Ignore optional terms in Release Profiles
Terms marked "optional" (preferred, ignored, and required) are now
separated out from the main list of terms. Right now, any optional terms
are NOT uploaded to Sonarr. In the future, I plan to add ways to
explicitly include optional terms.

The structure for optional terms in the guide is similar to that of
categories. A header OR line within the header section can mention the
word "optional" and that means any code blocks past that point until the
end of the section are treated as optional.

Other changes:

- Delete test trash.yml
- Add new ScopedState class

  Used to manage resetting certain parser state between different
  sections of the guide (for single code blocks or whole header
  sections)