feat(radarr): add custom formats by trash_id
There was an ambiguity with one custom format named 'DoVi'. This one had
two custom formats in the guide. The intent was for the user to choose
only one of these, but the name was kept identical so that name appeared
in the filename for media when it was renamed.

This allows the user to choose which of those two they want using the
`trash_id` property.