fix: Missing terms when using term filters
When using filters like `exclude`, it was possible for terms to not get
synced when they should have. This was due to a misunderstanding of how
`ExceptBy()` and `IntersectBy()` work. According to [an issue][1] on the
dotnet runtime repo, this is by design. The fix is to just avoid those
in favor of `Where()`.

Fixes #69.

[1]: https://github.com/dotnet/dotnet-api-docs/issues/7656