refactor: Delay file logging until IAppPaths initialized
The logger, which also writes to a file in addition to console,
requires `IAppPaths` in order to find the directory to place the log
files. However, this cannot be obtained until the system calculates
the app data directory OR the user specifies it with the `--app-data`
option.

A custom sink has been added that will allow the logger to write to
console without file logs until that initialization is performed and
the log directory is available.