fix: Use a less severe and better worded warning msg for CF scores
If scores are missing from a CF in either the guide or in the YAML
config, that CF is still synced but no score will be set in any quality
profiles.

The warning message for this was a bit misleading. It made users think
the CF itself would not be synced. The CF is always synced.

The message has been reworded to make this more clear and it is
downgraded from Warning to Informational.