fix: Fix false-positive duplicate score warnings
When doing a `sync --preview`, new custom formats are not created and
thus they never get an ID greater than `0`. Because of this, a
dictionary that tracks duplicates based on ID would result in warnings
about duplicate scores that made no sense.

We now index by Trash ID instead of Format ID, which is more accurate.