chore: tar.xz archives now set UID and GID to 0
The reason for this is to avoid the need for `--no-same-owner` when
extracting the archive as a super user (e.g. `sudo`). Without this
change, the UID and GID of `recyclarr` after it is extracted is whatever
user was used in github actions to create the archive, which isn't
really useful.

This change does not affect extractions as normal users; by default, the
UID/GID inside the archive is not used in those cases.