ci: Fix error while re-uploading signed artifacts
The Apple Signing step grabs an existing artifact, signs the binaries,
and then re-uploads the artifact, replacing the original one. The
upgrade to v4 of the `actions/upload-artifact` action broke this
behavior due to a breaking change documented in [the release notes][1]:

> Due to how Artifacts are created in this new version, it is no longer
> possible to upload to the same named Artifact multiple times. You must
> either split the uploads into multiple Artifacts with different names,
> or only upload once. Otherwise you will encounter an error.

The fix is to use the `overwrite: true` setting to force the old
behavior of replacing the previous (unsigned) artifact.

[1]: https://github.com/actions/upload-artifact/blob/main/README.md#breaking-changes