chore(docker): Build Recyclarr inside Dockerfile
For easier integration into docker buildx, Recyclarr is now built inside
of the Dockerfile as a dedicated build stage. This simplified the CI
workflows and local testing a bit.

The changes for this new process were driven by Microsoft's recommended
approach to building .NET applications inside Docker, which is
documented here:

https://github.com/dotnet/dotnet-docker/blob/main/samples/README.md