ci: Do not publish prerelease docker tags anymore
The `-dev.` docker tags create a lot of noise and cloud out the actual
non-prerelease versions in the tag listings on Docker Hub and GHCR.
Additionally, there's no built-in retention policies in these registries
and manual attempts to clean up old dev images are difficult.

Users that are interested in immutable tags can pin the `edge` tag to a
digest.