wip: Refactor Spectre Console to use a provider for CancellationToken
Decided to back out of this because hard-coding services to use a
single, injected CancellationToken would make the system too rigid. The
original goal for using a service was to limit the amount of method
calls that need to propagate a CancellationToken, but this has a
negative impact on the ability to use CancellationToken objects in other
contexts, such as a GUI application.