ci: Add job to test artifact downloads
Downgrade the artifact upload/download actions to v3 since none of the
v4 versions seem to address the "Not a valid zip file" issue.

Modified the codesign job to additionally run on PRs, so that when
Renovate upgrades the upload/download artifacts, we can tell from the PR
if the ZIP file issue is fixed.

For more info, see:

- https://github.com/actions/download-artifact/issues/328
- https://github.com/docker/build-push-action/issues/1167