fix: Better change detection for CF specification fields
I've made some updates to Recyclarr's CF "field" handling for
specifications, specifically addressing the issues regarding processing
language specifications

Here's a quick rundown of how Recyclarr processes fields for a language
spec after this change:

When loading the following fields from the guide:

```json
"fields": {
  "value": 8,
  "exceptLanguage": false
}
```

Recyclarr transforms it into the API-compatible format:

```json
"fields": [
  {
    "name": "value",
    "value": 8
  },
  {
    "name": "exceptLanguage",
    "value": false
  }
]
```

Next, it retrieves CF data from the API. For instance, if the API
returns:

```json
"fields": [
  {
    "name": "value",
    "value": 10
  },
  {
    "name": "foo",
    "value": "bar"
  },
  {
    "name": "exceptLanguage",
    "value": true
  }
]
```

Recyclarr compares the two sets of fields by matching the `name`
attributes from the guide to those from the API. Any fields present in
the API but absent in the guide are ignored. The values for matching
fields are then updated accordingly, and these changes are pushed back.
For this example:

- The field `value` is updated from `8` to `10`.
- The field `exceptLanguage` is updated from `false` to `true`.
- The field `foo` is ignored since there's no corresponding field in the
  guide.